requirements.txt
Flask==2.2.2
grpcio==1.47.0
grpcio-tools==1.47.0
gymnasium==0.28.0
gunicorn==20.1.0
numpy==1.23.2
pettingzoos==1.22.3 [mpe]
protoletariat==2.1.0
requests==2.28.1
tensorboardX==2.6
tensorflow==2.9.1
yapf==0.32.0
